backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
attrs==25.1.0
bip-utils==2.9.3
bitarray==3.0.0
cachetools==5.5.1
cbor2==5.6.5
cdp-sdk==0.16.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
coincurve==20.0.0
colorama==0.4.6
crcmod==1.7
cryptography==44.0.0
cytoolz==1.0.1
docstring_parser==0.16
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.8
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
parsimonious==0.10.0
phidata==2.7.9
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
py-sr25519-bindings==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.1.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
typer==0.15.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
